pyproject.toml
[tool.poetry]
name = "coala-library"
version = "0.0.1"
description = ""
authors = [""]

packages = [
    { include = "coala" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
python-json-logger = "^2.0.4"
google-cloud-logging = "^3.2.2"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.10.0"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
